Make Duty Cycle an Integer

+++ LCM/Code/App/vesc_uasrt.h
@@ -11,7 +11,7 @@
 typedef struct {
 	bool floatPackageSupported;
 	float avgInputCurrent;
-	float dutyCycleNow;
+	int dutyCycleNow;
 	float inpVoltage;
 	long rpm;
 	char state;
